# cpu, conv: fix L2 blocking for AVX2