@georgen117 2025-09-29
Closing due to misunderstanding the performance data spreadsheet.